savegame: move common logic to TRX
This moves the bulk of savegame logic to TRX.
