+++ src/libtrx/include/libtrx/game/lara/misc.h
@@ -7,3 +7,4 @@ void Lara_Extinguish(void);
 
 int16_t Lara_FloorFront(const ITEM *item, int16_t ang, int32_t dist);
 void Lara_GetCollisionInfo(const ITEM *item, COLL_INFO *coll);
+extern void Lara_CatchFire(void);
