+++ src/libtrx/include/libtrx/game/pathing/lot.h
@@ -5,3 +5,4 @@
 extern bool LOT_EnableBaddieAI(int16_t item_num, bool always);
 extern void LOT_DisableBaddieAI(int16_t item_num);
 extern CREATURE *LOT_GetBaddieSlot(int32_t i);
+extern void LOT_ClearLOT(LOT_INFO *LOT);
