+++ src/libtrx/include/libtrx/game/savegame/const.h
@@ -1,4 +1,3 @@
 #pragma once
 
 #define SAVEGAME_CURRENT_VERSION 8
-#define MAX_STRATEGIES 2
