+++ src/tr1/game/lara/misc.h
@@ -23,4 +23,3 @@ int32_t Lara_GetWaterDepth(int32_t x, int32_t y, int32_t z, int16_t room_num);
 void Lara_TestWaterDepth(ITEM *item, const COLL_INFO *coll);
 bool Lara_TestWaterStepOut(ITEM *item, const COLL_INFO *coll);
 bool Lara_TestWaterClimbOut(ITEM *item, COLL_INFO *coll);
-void Lara_CatchFire(void);
