+++ src/tr1/game/lot.h
@@ -10,4 +10,3 @@ void LOT_InitialiseArray(void);
 void LOT_InitialiseSlot(int16_t item_num, int32_t slot);
 void LOT_CreateZone(ITEM *item);
 void LOT_InitialiseLOT(LOT_INFO *LOT);
-void LOT_ClearLOT(LOT_INFO *LOT);
