+++ src/tr1/game/savegame/savegame.c
@@ -3,272 +3,21 @@
 #include "game/game.h"
 #include "game/game_flow.h"
 #include "game/game_string.h"
-#include "game/inventory.h"
-#include "game/items.h"
-#include "game/lot.h"
-#include "game/objects/vars.h"
 #include "game/requester.h"
-#include "game/room.h"
-#include "global/const.h"
-#include "global/types.h"
 #include "global/vars.h"
 
-#include <libtrx/benchmark.h>
 #include <libtrx/config.h>
 #include <libtrx/debug.h>
-#include <libtrx/enum_map.h>
-#include <libtrx/filesystem.h>
 #include <libtrx/game/gun/const.h>
-#include <libtrx/game/objects/traps/movable_block.h>
-#include <libtrx/memory.h>
-#include <libtrx/strings.h>
 
-#include <stdint.h>
-#include <stdio.h>
-#include <string.h>
-
-#define SAVES_DIR "saves"
-
-static int32_t m_SaveSlots = 0;
-static int16_t m_NewestSlot = -1;
-static int32_t m_SaveCounter = 0;
-static int32_t m_SavedGames = 0;
-static SAVEGAME_INFO *m_SavegameInfo = nullptr;
-static RESUME_INFO *m_ResumeInfo = nullptr;
-
-static int32_t m_StrategyCount = 0;
-static SAVEGAME_STRATEGY m_Strategies[MAX_STRATEGIES];
-
-static void M_ClearSlots(void);
-static bool M_FillSlot(
-    const SAVEGAME_STRATEGY strategy, int32_t slot_num, const char *path);
-static void M_ScanSavedGamesDir(const char *dir_path);
-static void M_LoadPreprocess(void);
-static void M_LoadPostprocess(void);
-
-static void M_ClearSlots(void)
+int32_t Savegame_GetSlotCount(void)
 {
-    if (m_SavegameInfo == nullptr) {
-        return;
-    }
-
-    for (int32_t i = 0; i < m_SaveSlots; i++) {
-        SAVEGAME_INFO *const savegame_info = &m_SavegameInfo[i];
-        savegame_info->format = 0;
-        savegame_info->counter = -1;
-        savegame_info->level_num = -1;
-        Memory_FreePointer(&savegame_info->full_path);
-        Memory_FreePointer(&savegame_info->level_title);
-    }
+    return g_Config.gameplay.maximum_save_slots;
 }
 
-static bool M_FillSlot(
-    const SAVEGAME_STRATEGY strategy, const int32_t slot_num,
-    const char *const path)
+void Savegame_HighlightNewestSlot(void)
 {
-    SAVEGAME_INFO *const savegame_info = &m_SavegameInfo[slot_num];
-    if (strategy.format <= savegame_info->format) {
-        // do not override already filled slots or for less preferred strategies
-        return true;
-    }
-
-    bool result = false;
-    MYFILE *const fp = File_Open(path, FILE_OPEN_READ);
-    if (fp != nullptr) {
-        if (strategy.fill_info_func(fp, savegame_info)) {
-            savegame_info->format = strategy.format;
-            Memory_FreePointer(&savegame_info->full_path);
-            savegame_info->full_path = Memory_DupStr(path);
-            result = true;
-        }
-        File_Close(fp);
-    }
-    return result;
-}
-
-static void M_ScanSavedGamesDir(const char *const dir_path)
-{
-    void *dir_handle = File_OpenDirectory(dir_path);
-    if (dir_handle == nullptr) {
-        return;
-    }
-
-    while (true) {
-        const char *file_name = File_ReadDirectory(dir_handle);
-        if (file_name == nullptr) {
-            break;
-        }
-        if (strcmp(file_name, ".") == 0 || strcmp(file_name, "..") == 0) {
-            continue;
-        }
-
-        for (int32_t i = 0; i < m_StrategyCount; i++) {
-            const SAVEGAME_STRATEGY strategy = m_Strategies[i];
-            if (!strategy.allow_load) {
-                continue;
-            }
-
-            int32_t slot = -1;
-            const int32_t parsed =
-                sscanf(file_name, strategy.get_save_file_pattern_func(), &slot);
-            if (parsed == 1 && slot >= 0 && slot < m_SaveSlots) {
-                char *file_path = String_Format("%s/%s", dir_path, file_name);
-                M_FillSlot(strategy, slot, file_path);
-                Memory_FreePointer(&file_path);
-            }
-        }
-    }
-
-    File_CloseDirectory(dir_handle);
-}
-
-static void M_LoadPreprocess(void)
-{
-    Savegame_InitCurrentInfo();
-}
-
-static void M_LoadPostprocess(void)
-{
-    for (int32_t i = 0; i < Item_GetLevelCount(); i++) {
-        ITEM *const item = Item_Get(i);
-        const OBJECT *const obj = Object_Get(item->object_id);
-
-        if (obj->save_position && obj->shadow_size) {
-            int16_t room_num = item->room_num;
-            const SECTOR *const sector = Room_GetSector(
-                item->pos.x, item->pos.y, item->pos.z, &room_num);
-            item->floor =
-                Room_GetHeight(sector, item->pos.x, item->pos.y, item->pos.z);
-        }
-
-        if (obj->save_flags != 0) {
-            item->flags &= 0xFF00;
-        }
-
-        if (obj->handle_save_func != nullptr) {
-            obj->handle_save_func(item, SAVEGAME_STAGE_AFTER_LOAD);
-        }
-    }
-
-    MovableBlock_SetupFloor();
-
-    if (Game_GetBonusFlag() != GBF_NONE) {
-        g_Config.profile.new_game_plus_unlock = true;
-    }
-
-    LOT_ClearLOT(&g_Lara.lot);
-}
-
-void Savegame_RegisterStrategy(const SAVEGAME_STRATEGY strategy)
-{
-    ASSERT(m_StrategyCount < MAX_STRATEGIES);
-    m_Strategies[m_StrategyCount] = strategy;
-    m_StrategyCount++;
-}
-
-void Savegame_Init(void)
-{
-    m_ResumeInfo = Memory_Alloc(
-        sizeof(RESUME_INFO)
-        * (GF_GetLevelTable(GFLT_MAIN)->count
-           + (GF_GetLevelTable(GFLT_DEMOS)->count >= 0 ? 1 : 0)));
-    m_SaveSlots = g_Config.gameplay.maximum_save_slots;
-    m_SavegameInfo = Memory_Alloc(sizeof(SAVEGAME_INFO) * m_SaveSlots);
-
-    const GF_LEVEL_TABLE *const level_table = GF_GetLevelTable(GFLT_DEMOS);
-    for (int32_t i = 0; i < level_table->count; i++) {
-        RESUME_INFO *const resume_info =
-            Savegame_GetCurrentInfo(&level_table->levels[i]);
-        resume_info->flags.available = 1;
-        resume_info->flags.costume = 0;
-        resume_info->small_medipacks = 0;
-        resume_info->large_medipacks = 0;
-        resume_info->num_scions = 0;
-        resume_info->flags.has_pistols = 1;
-        resume_info->flags.has_shotgun = 0;
-        resume_info->flags.has_magnums = 0;
-        resume_info->flags.has_uzis = 0;
-        resume_info->pistol_ammo = 1000;
-        resume_info->shotgun_ammo = 0;
-        resume_info->magnum_ammo = 0;
-        resume_info->uzi_ammo = 0;
-        resume_info->gun_status = LGS_ARMLESS;
-        resume_info->equipped_gun_type = LGT_PISTOLS;
-        resume_info->holsters_gun_type = LGT_PISTOLS;
-        resume_info->back_gun_type = LGT_UNARMED;
-        resume_info->lara_hitpoints = LARA_MAX_HITPOINTS;
-    }
-}
-
-RESUME_INFO *Savegame_GetCurrentInfo(const GF_LEVEL *const level)
-{
-    ASSERT(m_ResumeInfo != nullptr);
-    ASSERT(level != nullptr);
-    if (GF_GetLevelTableType(level->type) == GFLT_MAIN) {
-        return &m_ResumeInfo[level->num];
-    } else if (level->type == GFL_DEMO) {
-        return &m_ResumeInfo[GF_GetLevelTable(GFLT_MAIN)->count];
-    }
-    LOG_WARNING(
-        "Warning: unable to get resume info for level %d (type=%s)", level->num,
-        ENUM_MAP_TO_STRING(GF_LEVEL_TYPE, level->type));
-    return nullptr;
-}
-
-void Savegame_SetCurrentInfo(const int32_t current_slot, const int32_t src_slot)
-{
-    m_ResumeInfo[current_slot] = m_ResumeInfo[src_slot];
-}
-
-void Savegame_Shutdown(void)
-{
-    M_ClearSlots();
-    Memory_FreePointer(&m_SavegameInfo);
-    Memory_FreePointer(&m_ResumeInfo);
-}
-
-bool Savegame_IsInitialised(void)
-{
-    return m_SavegameInfo != nullptr;
-}
-
-void Savegame_ProcessItemsBeforeLoad(void)
-{
-    for (int32_t i = 0; i < Item_GetLevelCount(); i++) {
-        ITEM *const item = Item_Get(i);
-        const OBJECT *const obj = Object_Get(item->object_id);
-        if (obj->handle_save_func != nullptr) {
-            obj->handle_save_func(item, SAVEGAME_STAGE_BEFORE_LOAD);
-        }
-    }
-}
-
-void Savegame_ProcessItemsBeforeSave(void)
-{
-    for (int32_t i = 0; i < Item_GetLevelCount(); i++) {
-        ITEM *const item = Item_Get(i);
-        const OBJECT *const obj = Object_Get(item->object_id);
-        if (obj->handle_save_func != nullptr) {
-            obj->handle_save_func(item, SAVEGAME_STAGE_BEFORE_SAVE);
-        }
-    }
-}
-
-void Savegame_InitCurrentInfo(void)
-{
-    const GF_LEVEL_TABLE *const level_table = GF_GetLevelTable(GFLT_MAIN);
-    for (int32_t i = 0; i < level_table->count; i++) {
-        const GF_LEVEL *const level = &level_table->levels[i];
-        Savegame_ResetCurrentInfo(level);
-        Savegame_ApplyLogicToCurrentInfo(level);
-        Savegame_GetCurrentInfo(level)->flags.available = 0;
-    }
-    if (GF_GetGymLevel() != nullptr) {
-        Savegame_GetCurrentInfo(GF_GetGymLevel())->flags.available = 1;
-    }
-    if (GF_GetFirstLevel() != nullptr) {
-        Savegame_GetCurrentInfo(GF_GetFirstLevel())->flags.available = 1;
-    }
+    g_SavegameRequester.requested = MAX(0, Savegame_GetHighestSlot());
 }
 
 void Savegame_ApplyLogicToCurrentInfo(const GF_LEVEL *const level)
@@ -368,266 +117,13 @@ void Savegame_ApplyLogicToCurrentInfo(const GF_LEVEL *const level)
     }
 }
 
-void Savegame_ResetCurrentInfo(const GF_LEVEL *const level)
-{
-    LOG_INFO("Resetting resume info for level #%d", level->num);
-    RESUME_INFO *const current = Savegame_GetCurrentInfo(level);
-    memset(current, 0, sizeof(RESUME_INFO));
-}
-
-void Savegame_CarryCurrentInfoToNextLevel(
-    const GF_LEVEL *const src_level, const GF_LEVEL *const dst_level)
-{
-    LOG_INFO(
-        "Copying resume info from level #%d to level #%d", src_level->num,
-        dst_level->num);
-    RESUME_INFO *const src_resume = Savegame_GetCurrentInfo(src_level);
-    RESUME_INFO *const dst_resume = Savegame_GetCurrentInfo(dst_level);
-    memcpy(dst_resume, src_resume, sizeof(RESUME_INFO));
-}
-
-void Savegame_PersistGameToCurrentInfo(const GF_LEVEL *const level)
-{
-    ASSERT(level != nullptr);
-    RESUME_INFO *current = Savegame_GetCurrentInfo(level);
-
-    current->lara_hitpoints = g_LaraItem->hit_points;
-    current->flags.available = 1;
-
-    current->pistol_ammo = 1000;
-    if (Inv_RequestItem(O_PISTOL_ITEM)) {
-        current->flags.has_pistols = 1;
-    } else {
-        current->flags.has_pistols = 0;
-    }
-
-    if (Inv_RequestItem(O_MAGNUM_ITEM)) {
-        current->magnum_ammo = g_Lara.magnum_ammo.ammo;
-        current->flags.has_magnums = 1;
-    } else {
-        current->magnum_ammo =
-            Inv_RequestItem(O_MAGNUM_AMMO_ITEM) * MAGNUM_AMMO_QTY;
-        current->flags.has_magnums = 0;
-    }
-
-    if (Inv_RequestItem(O_UZI_ITEM)) {
-        current->uzi_ammo = g_Lara.uzi_ammo.ammo;
-        current->flags.has_uzis = 1;
-    } else {
-        current->uzi_ammo = Inv_RequestItem(O_UZI_AMMO_ITEM) * UZI_AMMO_QTY;
-        current->flags.has_uzis = 0;
-    }
-
-    if (Inv_RequestItem(O_SHOTGUN_ITEM)) {
-        current->shotgun_ammo = g_Lara.shotgun_ammo.ammo;
-        current->flags.has_shotgun = 1;
-    } else {
-        current->shotgun_ammo =
-            Inv_RequestItem(O_SHOTGUN_AMMO_ITEM) * SHOTGUN_AMMO_QTY;
-        current->flags.has_shotgun = 0;
-    }
-
-    current->small_medipacks = Inv_RequestItem(O_SMALL_MEDIPACK_ITEM);
-    current->large_medipacks = Inv_RequestItem(O_LARGE_MEDIPACK_ITEM);
-    current->num_scions = Inv_RequestItem(O_SCION_ITEM_1);
-
-    current->equipped_gun_type = g_Lara.gun_type;
-    current->holsters_gun_type = g_Lara.holsters_gun_type;
-    current->back_gun_type = g_Lara.back_gun_type;
-    if (g_Lara.gun_status == LGS_READY) {
-        current->gun_status = LGS_READY;
-    } else {
-        current->gun_status = LGS_ARMLESS;
-    }
-}
-
-int32_t Savegame_GetLevelNumber(const int32_t slot_num)
-{
-    return m_SavegameInfo[slot_num].level_num;
-}
-
-int32_t Savegame_GetSlotCount(void)
-{
-    return m_SaveSlots;
-}
-
-bool Savegame_IsSlotFree(const int32_t slot_num)
-{
-    return m_SavegameInfo[slot_num].level_num == -1;
-}
-
-bool Savegame_Load(const int32_t slot_num)
-{
-    const SAVEGAME_INFO *const savegame_info = &m_SavegameInfo[slot_num];
-    ASSERT(savegame_info->format != SAVEGAME_FORMAT_INVALID);
-
-    M_LoadPreprocess();
-
-    bool ret = false;
-    for (int32_t i = 0; i < m_StrategyCount; i++) {
-        const SAVEGAME_STRATEGY strategy = m_Strategies[i];
-        if (savegame_info->format == strategy.format) {
-            MYFILE *const fp =
-                File_Open(savegame_info->full_path, FILE_OPEN_READ);
-            if (fp != nullptr) {
-                ret = strategy.load_from_file_func(fp);
-                File_Close(fp);
-            }
-            break;
-        }
-    }
-
-    if (ret) {
-        M_LoadPostprocess();
-    }
-
-    Savegame_SetInitialVersion(m_SavegameInfo[slot_num].initial_version);
-    return ret;
-}
-
-bool Savegame_Save(const int32_t slot_num)
-{
-    bool ret = false;
-    Savegame_BindSlot(slot_num);
-
-    File_CreateDirectory(SAVES_DIR);
-
-    const GF_LEVEL *const current_level = Game_GetCurrentLevel();
-    Savegame_PersistGameToCurrentInfo(current_level);
-
-    const GF_LEVEL_TABLE *const level_table = GF_GetLevelTable(GFLT_MAIN);
-    for (int32_t i = 0; i < level_table->count; i++) {
-        const GF_LEVEL *const level = &level_table->levels[i];
-        if (level->type == GFL_CURRENT) {
-            Savegame_SetCurrentInfo(i, current_level->num);
-        }
-    }
-    const char *const level_title = Game_GetCurrentLevel()->title;
-
-    SAVEGAME_INFO *const savegame_info = &m_SavegameInfo[slot_num];
-    const bool was_slot_empty = savegame_info->full_path == nullptr;
-
-    m_SaveCounter++;
-    for (int32_t i = 0; i < m_StrategyCount; i++) {
-        const SAVEGAME_STRATEGY strategy = m_Strategies[i];
-        if (!strategy.allow_save || strategy.save_to_file_func == nullptr) {
-            continue;
-        }
-
-        char *filename =
-            String_Format(strategy.get_save_file_pattern_func(), slot_num);
-        char *full_path = String_Format("%s/%s", SAVES_DIR, filename);
-
-        MYFILE *const fp = File_Open(full_path, FILE_OPEN_WRITE);
-        if (fp != nullptr) {
-            strategy.save_to_file_func(fp, savegame_info);
-            savegame_info->format = strategy.format;
-            Memory_FreePointer(&savegame_info->full_path);
-            savegame_info->full_path = Memory_DupStr(File_GetPath(fp));
-            savegame_info->counter = m_SaveCounter;
-            savegame_info->level_num = current_level->num;
-            savegame_info->level_title =
-                level_title != nullptr ? Memory_DupStr(level_title) : nullptr;
-            File_Close(fp);
-            ret = true;
-        }
-
-        Memory_FreePointer(&filename);
-        Memory_FreePointer(&full_path);
-    }
-
-    if (ret) {
-        m_NewestSlot = slot_num;
-        if (was_slot_empty) {
-            m_SavedGames++;
-        }
-        Savegame_HighlightNewestSlot();
-    } else {
-        m_SaveCounter--;
-    }
-
-    return ret;
-}
-
-bool Savegame_UpdateDeathCounters(
-    const int32_t slot_num, const int32_t death_count)
-{
-    ASSERT(slot_num >= 0);
-    const SAVEGAME_INFO *const savegame_info = &m_SavegameInfo[slot_num];
-    ASSERT(savegame_info->format != SAVEGAME_FORMAT_INVALID);
-
-    bool ret = false;
-    for (int32_t i = 0; i < m_StrategyCount; i++) {
-        const SAVEGAME_STRATEGY strategy = m_Strategies[i];
-        if (savegame_info->format == strategy.format
-            && strategy.update_death_counters_func != nullptr) {
-            MYFILE *const fp =
-                File_Open(savegame_info->full_path, FILE_OPEN_READ_WRITE);
-            if (fp != nullptr) {
-                ret = strategy.update_death_counters_func(fp, death_count);
-                File_Close(fp);
-            }
-            break;
-        }
-    }
-    return ret;
-}
-
-bool Savegame_LoadOnlyResumeInfo(int32_t slot_num)
-{
-    const SAVEGAME_INFO *const savegame_info = &m_SavegameInfo[slot_num];
-    ASSERT(savegame_info->format != SAVEGAME_FORMAT_INVALID);
-
-    bool ret = false;
-    for (int32_t i = 0; i < m_StrategyCount; i++) {
-        const SAVEGAME_STRATEGY strategy = m_Strategies[i];
-        if (savegame_info->format == strategy.format) {
-            MYFILE *const fp =
-                File_Open(savegame_info->full_path, FILE_OPEN_READ);
-            if (fp != nullptr) {
-                ret = strategy.load_only_resume_info_func(fp);
-                File_Close(fp);
-            }
-            break;
-        }
-    }
-
-    Savegame_SetInitialVersion(m_SavegameInfo[slot_num].initial_version);
-    return ret;
-}
-
-void Savegame_ScanSavedGames(void)
-{
-    BENCHMARK benchmark = Benchmark_Start();
-    M_ClearSlots();
-
-    m_SaveCounter = 0;
-    m_SavedGames = 0;
-    m_NewestSlot = -1;
-
-    M_ScanSavedGamesDir(SAVES_DIR);
-    M_ScanSavedGamesDir(".");
-
-    for (int32_t i = 0; i < m_SaveSlots; i++) {
-        SAVEGAME_INFO *savegame_info = &m_SavegameInfo[i];
-        if (savegame_info->level_title != nullptr) {
-            if (savegame_info->counter > m_SaveCounter) {
-                m_SaveCounter = savegame_info->counter;
-                m_NewestSlot = i;
-            }
-            m_SavedGames++;
-        }
-    }
-    Benchmark_End(&benchmark, nullptr);
-}
-
-void Savegame_FillAvailableSaves(REQUEST_INFO *req)
+void Savegame_FillAvailableSaves(REQUEST_INFO *const req)
 {
     Requester_ClearTextstrings(req);
     Requester_Init(req, Savegame_GetSlotCount());
 
-    for (int i = 0; i < req->max_items; i++) {
-        SAVEGAME_INFO *savegame_info = &m_SavegameInfo[i];
+    for (int32_t i = 0; i < req->max_items; i++) {
+        const SAVEGAME_INFO *const savegame_info = Savegame_GetSavegameInfo(i);
 
         if (savegame_info->level_title != nullptr) {
             Requester_AddItem(
@@ -645,7 +141,7 @@ void Savegame_FillAvailableSaves(REQUEST_INFO *req)
     }
 }
 
-void Savegame_FillAvailableLevels(REQUEST_INFO *req)
+void Savegame_FillAvailableLevels(REQUEST_INFO *const req)
 {
     ASSERT(req != nullptr);
     const int32_t slot_num = g_GameInfo.select_save_slot;
@@ -653,7 +149,8 @@ void Savegame_FillAvailableLevels(REQUEST_INFO *req)
         return;
     }
 
-    const SAVEGAME_INFO *const savegame_info = &m_SavegameInfo[slot_num];
+    const SAVEGAME_INFO *const savegame_info =
+        Savegame_GetSavegameInfo(slot_num);
     if (!savegame_info->features.select_level) {
         Requester_AddItem(req, true, "%s", GS(PASSPORT_LEGACY_SELECT_LEVEL_1));
         Requester_AddItem(req, true, "%s", GS(PASSPORT_LEGACY_SELECT_LEVEL_2));
@@ -677,29 +174,4 @@ void Savegame_FillAvailableLevels(REQUEST_INFO *req)
 
     req->requested = 0;
     req->line_offset = 0;
-}
-
-void Savegame_HighlightNewestSlot(void)
-{
-    g_SavegameRequester.requested = MAX(0, m_NewestSlot);
-}
-
-bool Savegame_RestartAvailable(int32_t slot_num)
-{
-    if (slot_num == -1) {
-        return true;
-    }
-
-    SAVEGAME_INFO *savegame_info = &m_SavegameInfo[slot_num];
-    return savegame_info->features.restart;
-}
-
-int32_t Savegame_GetCounter(void)
-{
-    return m_SaveCounter;
-}
-
-int32_t Savegame_GetTotalCount(void)
-{
-    return m_SavedGames;
 }
