+++ src/tr2/game/lot.h
@@ -7,4 +7,3 @@
 void LOT_InitialiseArray(void);
 void LOT_InitialiseSlot(int16_t item_num, int32_t slot);
 void LOT_CreateZone(ITEM *item);
-void LOT_ClearLOT(LOT_INFO *LOT);
